refactor oai-compatible code to reuse more code
