+++ agents/OpenAI_Compatible_API_Agent/Python/openai_compatible_api.py
@@ -1,20 +1,16 @@
-import asyncio
-import json
-import time
 from pathlib import Path
-from typing import Optional, List
-import tiktoken
-from pydantic import BaseModel
+
 from starlette.responses import StreamingResponse
 
 from fastapi import FastAPI, Request
 from threading import local
 
+from agents.OpenAI_Compatible_API_Agent.Python.open_ai_helper import ChatInstance, \
+    ChatCompletionRequest, _resp_sync, _resp_async_generator, get_models
 from .pgpt_api import PrivateGPTAPI
 
-from ...AgentInterface.Python.agent import PrivateGPTAgent
 from ...AgentInterface.Python.config import Config, ConfigError
-
+import uvicorn
 
 app = FastAPI(title="OpenAI-compatible API for PrivateGPT")
 request_context = local()
@@ -28,24 +24,6 @@ except ConfigError as e:
     print(f"Configuration Error: {e}")
     exit(1)
 
-class ChatInstance:
-    def __init__(self, api_key: str, agent: PrivateGPTAgent | PrivateGPTAPI):
-        self.api_key = api_key
-        self.agent = agent
-
-
-# data models
-class Message(BaseModel):
-    role: str
-    content: str
-
-
-class ChatCompletionRequest(BaseModel):
-    model: Optional[str] = "PGPT - Mistral NeMo 12B"
-    messages: List[Message]
-    max_tokens: Optional[int] = 2048
-    temperature: Optional[float] = 0.1
-    stream: Optional[bool] = False
 
 
 @app.middleware("http")
@@ -100,92 +78,14 @@ async def chat_completions(request: ChatCompletionRequest):
         return _resp_sync(response, request)
 
 
-def _resp_sync(response: json,request):
-    user_input = ""
-    for message in request.messages:
-        user_input += json.dumps({'role': message.role, 'content': message.content})
-
-    num_tokens_request = num_tokens_from_string(user_input, "o200k_base")
-    num_tokens_reply = num_tokens_from_string(response["answer"], "o200k_base")
-    num_tokens_overall = num_tokens_request + num_tokens_reply
-    citations = []
-    if "sources" in response:
-        citations = response["sources"]
-
-    return {
-        "id": response["chatId"],
-        "object": "chat.completion",
-        "created": time.time(),
-        "model": request.model,
-        "choices": [{"message": Message(role="assistant", content=response["answer"])}],
-        "citations": citations,
-        "usage": {
-            "prompt_tokens": num_tokens_request,
-            "completion_tokens": num_tokens_reply,
-            "total_tokens": num_tokens_overall
-        }
-    }
-
-async def _resp_async_generator(response: json, request):
-    #print("RESPONSE: " + str(response))
-    # Count tokens
-    user_input = ""
-    for message in request.messages:
-        user_input += json.dumps({'role': message.role, 'content': message.content})
-
-    num_tokens_request = num_tokens_from_string(user_input, "o200k_base")
-    num_tokens_reply = num_tokens_from_string(response["answer"], "o200k_base")
-    num_tokens_overall = num_tokens_request + num_tokens_reply
-
-    tokens = response["answer"].split(" ")
-    citations = []
-    if "sources" in response:
-        citations = response["sources"]
-
-    for i, token in enumerate(tokens):
-        chunk = {
-            "id": i,
-            "object": "chat.completion.chunk",
-            "created": time.time(),
-            "model": request.model,
-            "choices": [{"delta": {"content": token + " "}}],
-            "citations": citations,
-            "usage": {
-                "prompt_tokens": num_tokens_request,
-                "completion_tokens": num_tokens_reply,
-                "total_tokens": num_tokens_overall
-            }
-        }
-        yield f"data: {json.dumps(chunk)}\n\n"
-        await asyncio.sleep(0.05)
-    yield "data: [DONE]\n\n"
-
-def num_tokens_from_string(string: str, encoding_name: str) -> int:
-    """Returns the number of tokens in a text string."""
-    encoding = tiktoken.get_encoding(encoding_name)
-    num_tokens = len(encoding.encode(string))
-    return num_tokens
-
-
 
 @app.get("/models")
-def get_models():
-    return {
-        "object": "list",
-        "data": [
-            {
-                "id": "pgpt-mistral-nemo-12b",
-                "object": "model",
-                "created": 1686935002,
-                "owned_by": "Fujitsu"
-            }
-        ]
-    }
+def return_models():
+    return get_models()
 
 
 
 if __name__ == "__main__":
-    import uvicorn
     api_ip = config.get("api_ip", "0.0.0.0")
     api_port = config.get("api_port", 8002)
     uvicorn.run(app, host=api_ip, port=int(api_port))
